Split big test_check_meta.py into other files
