+++ python-scripts/metadatavalidator/tests/unit/checks/test_check_meta_task.py
@@ -1,0 +1,69 @@
+from lxml import etree
+import pytest
+
+from _utils import appendnode, D, xmlid
+
+from metadatavalidator.common import NAMESPACES
+from metadatavalidator.checks.check_meta import (
+    check_meta_category,
+    check_meta_task,
+)
+from metadatavalidator.exceptions import InvalidValueError
+
+
+def test_meta_task(tree):
+#     xmlcontent = """<article xmlns="http://docbook.org/ns/docbook" version="5.2">
+#     <info>
+#         <title>Test</title>
+#         <meta name="task">
+#             <phrase>Configuration</phrase>
+#         </meta>
+#     </info>
+# </article>"""
+    meta = D("meta", {"name": "task"},
+             D("phrase", "Configuration"))
+    appendnode(tree, meta)
+
+    config = dict(metadata=dict(require_meta_task=True,
+                                valid_meta_task=["Configuration"]))
+    assert check_meta_task(tree, config) is None
+
+
+def test_missing_child_meta_task(tree):
+#     xmlcontent = """<article xmlns="http://docbook.org/ns/docbook" version="5.2">
+#     <info>
+#         <title>Test</title>
+#         <meta name="task"/>
+#     </info>
+# </article>"""
+    meta = D("meta", {"name": "task"})
+    appendnode(tree, meta)
+
+    config = dict(metadata=dict(require_meta_task=True),
+                                valid_meta_task=["Configuration"])
+    with pytest.raises(InvalidValueError,
+                       match=r".*Couldn't find any child elements in meta.*"):
+        check_meta_task(tree, config)
+
+
+def test_duplicate_child_meta_task(tree):
+#     xmlcontent = """<article xmlns="http://docbook.org/ns/docbook" version="5.2">
+#     <info>
+#         <title>Test</title>
+#         <meta name="task">
+#             <phrase>Configuration</phrase>
+#             <phrase>Configuration</phrase>
+#         </meta>
+#     </info>
+# </article>"""
+    meta = D("meta", {"name": "task"},
+             D("phrase", "Configuration"),
+             D("phrase", "Configuration"),
+             )
+    appendnode(tree, meta)
+
+    config = dict(metadata=dict(require_meta_task=True),
+                                valid_meta_task=["Configuration"])
+    with pytest.raises(InvalidValueError,
+                       match=r".*Duplicate tasks found in meta.*"):
+        check_meta_task(tree, config)
